# docs/swagger.yaml
basePath: /api
definitions:
  controller.DatasetDTO:
    properties:
      created_at:
        type: string
      description:
        example: 数据集描述
        type: string
      entry_count:
        example: 100
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: 数据集名称
        type: string
      project_id:
        example: 1
        type: integer
      schema_definition:
        type: string
      storage_type:
        example: database
        type: string
      template_type:
        example: instruction_io
        type: string
      total_size:
        example: 1024
        type: integer
      updated_at:
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  controller.DatasetResponse:
    properties:
      data:
        $ref: '#/definitions/controller.DatasetDTO'
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  controller.DatasetsListData:
    properties:
      datasets:
        items:
          $ref: '#/definitions/controller.DatasetDTO'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  controller.DatasetsResponse:
    properties:
      data:
        $ref: '#/definitions/controller.DatasetsListData'
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  controller.ErrorResponse:
    properties:
      data: {}
      message:
        example: 发生错误
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  controller.SuccessResponse:
    properties:
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.Dataset:
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for MLcore-Engine.
  title: MLcore-Engine API
  version: "1.0"
paths:
  /api/dataset:
    get:
      consumes:
      - application/json
      description: 获取当前用户有权限访问的数据集列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetsResponse'
      summary: 获取数据集列表
      tags:
      - Dataset
    post:
      consumes:
      - application/json
      description: 创建新的数据集
      parameters:
      - description: 数据集信息
        in: body
        name: dataset
        required: true
        schema:
          $ref: '#/definitions/model.Dataset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 创建数据集
      tags:
      - Dataset
  /api/dataset/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的数据集
      parameters:
      - description: 数据集ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 删除数据集
      tags:
      - Dataset
    get:
      consumes:
      - application/json
      description: 获取单个数据集的详细信息
      parameters:
      - description: 数据集ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取数据集详情
      tags:
      - Dataset
    put:
      consumes:
      - application/json
      description: 更新数据集的基本信息
      parameters:
      - description: 数据集ID
        in: path
        name: id
        required: true
        type: integer
      - description: 数据集信息
        in: body
        name: dataset
        required: true
        schema:
          $ref: '#/definitions/model.Dataset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 更新数据集
      tags:
      - Dataset
...
  /api/dataset/project/{projectId}:
    get:
      consumes:
      - application/json
      description: 获取指定项目的所有数据集
      parameters:
      - description: 项目ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetsResponse'
      summary: 获取项目数据集
      tags:
      - Dataset
  /api/dataset/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索数据集
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DatasetsResponse'
      summary: 搜索数据集
      tags:
      - Dataset
...
swagger: "2.0"